Updated some code in module jsonfn
Changed the call in the function json_editor to the function json_pruner so that it passes a deep copy of the full startup JSON data, rather than a deep copy of the Items array.

Also updated the call in the function json_saver to the function generate_user_edited_data in the data_generate module. In the previous commit, this call was passing in "F" as a hardcoded strings. Refactored this to use the ItemTypeVals enum class member Full.

Tested that the call from json_pruner to generate_user_edited_data is passing in the correct data. Now, the rest of the code needs to be added to take the correctly updated JSON data and save it.