Started work on refactoring the function match_scenario
Removed the various existing cases and created a list called error_scenarios. The idea is to make it easier to figure out the new cases for each error message.

Also set up the first 4 cases by checking for the commonality for cases 1-3, then running a second match-case block for the last data_validation dictionary key. For the fourth case, the keys that matter are checked.

See changes for details.