Refactored the function match_scenario to properly account for the 4 valid scenarios and the 5 possible error scenarios
Based on changes to the commit-deets.txt file and the comments thread for issue Work on menu option: Delete a startup item #5, refactored the match_scenario function.

Tested that it works correctly to recognize all 4 valid scenarios and all 5 possible error scenarios. Currently, the function returns the correct data but also prints out extra information for confirmation that the function continues to work correctly.