Modified the json_editor function and the edit_startup_item function
While going over the code, realized that the function edit_startup_item in the module startup_edit, and the function json_editor in the module jsonfn, both needed slight refactoring.

The edit_startup_item function had listed in the docstring that a list was returned but the code was just returning the startup_item dictionary passed in as a parameter.

Refactored this to create a deep copy of the parameter to work with locally and then also to return a deep copy of the updated local startup item.

The json_editor function wasn't doing anything with the return value from the call to edit_startup_item because references were being passed back and forth. With the change to edit_startup_item to use deep copies and not references, refactored the call from json_editor so the actual startup item passed in is also updated using the return value.

The function json_editor doesn't do any deep copy but rather works solely with references to the startup data originally read in. While the call to the function json_pruner does pass in a deep copy, the call to the function json_saver only passed in a reference.

This may need to be refactored more so the calls from this json_editor function are all consistent as need to be, regardless of which functions are being called and which modules they are in.

Haven't tested these changes, but will as soon as possible.