Started coding in scenario 2 for the function generate_user_edited_data
Added the code to delete the requested startup item and to also update the Total Items property.

Also created an if and nested-if block to check if the existing startup items need to be update with respect to their Item Number property.

So far, for the situations where there's either only one startup item and it was deleted, or the deleted startup item was the last one in the Items array, the if-nested-if block does nothing.

Still need to code in the other 2 situations where the deleted startup item is the first in the Items array or it's in the middle of the Items array. Currently the if-nested-if block is set up to handle each situation separately, but the code for both might be the same.

To be clear, in all situations, the Total Items property is updated. It's just the Item Number property for existing startup items needs to be updated when appropriate.