Finished coding in scenario 2 for the function generate_user_edited_data
Following on the heels of the previous commit, coded in the situations where the deleted startup item requires the rest of the startup items to be updated for their Item Number property.

Realized while coding this is in that the if-nested-if block wasn't necessary and a single if-block was sufficient to catch these 2 situations.