Finished coding in the json_pruner function
Decided to pass in the startup JSON data but, within the function, a copy is made.

That copy is passed to the generate_user_edited_data function, which itself creates a new JSON object (Python dictionary) for the updated data.

The new object that's returned is then assigned to a variable in json_pruner and itself returned. Currently, the json_editor function, upon calling json_pruner, takes the returned and updated JSON data and assigns it to the json_data variable being used as a master reference.

Testing still needs to be done to see if this is sufficient and will allow the updated data to be saved. The menu that json_editor displays may also need to be updated so any startup items removed (or added) are reflected properly in the menu options.