Tested the changes and fixed any issues so the delete functionality now works
While testing, changed the startup item number counter to begin at 1 when a startup item is deleted and the other startup items have to be renumbered.

Moved the menu generation code in the function json_editor to be inside the loop and created a new boolean variable to determine when the menu needs to be recreated.

This allows for the menu to be created initially, when the user first chooses to edit the startup data, and whenever a startup item is deleted or added. This also allows for the menu to be displayed without needing to be recreated when the user has only edited an existing startup item's details.

Also, added some comments in the if-block in json_editor that deals with the user's menu option choice.