Adding a new JSON file called modified_startup_data.json file
This file is the result from testing for the previous commit. The specific testing done was the following against the default startup data:

1) Edited startup item 2 to remove an argument, change the startup item name, and change the startup item description.

2) Saved the modified startup item to disk.

3) Edited startup item 3 to add an argument, change the startup item name, and change the startup item description.

4) Deleted startup item 1, so default startup item 2 became item 1, and default startup item 3 became item 2.

5) Saved the modified full startup data to disk.

Between steps 4 and 5, I confirmed that the modified startup item saved to disk in step 2 was still correctly reflecting in the JSON file. I also confirmed that the startup item changes made in step 3 were reflecting in the menu options but not in the JSON file.

The new JSON file being added is the result of step 5. One edge case was found during testing, which I will recreate and save as another JSON file in a new commit.