Recreated the edge case mentioned in the previous commit
The testing was done against the default startup data and comprised the following steps:

1) Edited startup item 2 to remove an argument, change the startup item name, and change the startup item description.

2) Saved the modified startup item to disk. This was then copied to the JSON file "edge-case-one_startup_data.json".

3) Edited startup item 3 to add an argument, change the startup item name, and change the startup item description.

4) Deleted startup item 1, so default startup item 2 became item 1, and default startup item 3 became item 2.

5) Selected to edit startup item 2 from the udpated json_editor menu

6) Saved the startup item data to disk. This was then copied to the JSON file "edge-case-two_startup_data.json".

As can be seen, steps 1-4 are the same as in the previous commit. However, two new JSON files were created by copying the saved data at steps 2 and then 6.

This is because the JSON data saved at step 6 actually has 3 startup items, with items 1 and 3 being the exact same as the default startup data. However, since at step 5, the startup item selected was at that point startup item 2, in step 6, the default startup item 2 was replaced by this version of startup item 2.

In other words, if you compare the two edge case startup JSON files, you'll see the edge case happens when a startup item is modified but not saved, the user returns to the previous menu, deletes a different startup item but doesn't save the full startup data, then selects to edit the same startup item (which now has a different startup item number) and saves that item to disk.