Decided to move the function is_production from the main module to the helper module
The helper module is supposed to contain all helper functions - small functions that do one thing and don't really fit anywhere else.

However, the main module, comp_start, had a function called is_production that let the caller know if the current environment is a production one or not.

It did this by returning the value of a global boolean variable available in comp_start called is_prod. While this mechanism might change closer to release of this tool, for now this boolean variable has been manually set to False.

The is_production function was moved to the helper module as it should be part of that module for consistency.

The function was also modified slightly to directly return the variable value from the comp_start module.

All references to this function were updated as well.