Started changes necessary to refactor user_menu_chooser
Refactored the user_menu_chooser function definition to take in a list called menu_choices. Also removed the parameter second parameter total_menu_choices. Updated the docstring as well. However, haven't yet coded in the changes to the function.

Also went through and updated all calls to this function with the exception of the function edit_startup_item_arguments_list in the module startup_edit.

For all other calling functions, refactored things to create a list of strings and pass that in. Removed from the strings newline characters and numbering.