Refactored user_menu_chooser and tested the changes
Refactored the code in user_menu_chooser to build the menu using the new parameters.

Stepped through the changes to test both user_menu_chooser and the changes from the previous commit.

Made minor cleanup changes to the menu from the comp_start, jsonfn, and startup_edit modules. For the last module, cleaned up the menu in the function edit_startup_item.

The function edit_startup_item still needs to be refactored as it throws an error when it calls user_menu_chooser because it's using the old function definition.

Lastly, made a minor change to the way the JSON data is printerd out in the function prettify_json in the pretty module.