Modification to user_menu_chooser to use a local list
While doing further testing on user_menu_chooser, found a situation where the Quit Program option showed twice. Realized this was because Python passes references to lists and dictionaries.

For most of the menus, this doesn't cause any issues, as any menu choice that then leads to a sub-menu will have a reference to a different list passed in.

However, from the main menu, when option 1 is chosen to view what this tool is about, no new reference is passed, so the next call to user_menu_chooser uses the same list as the previous call.

However, this list has now been updated from the previous call to permanently have a Quit Program option, so another one is added the next time.

Fixed this by creating a local copy of the passed-in list in user_menu_chooser and adding the Quit Program option to that. That local copy is also used to build the menu, so now user_menu_chooser doesn't touch the list given to it regardless of the caller.