Created the function add_startup_item_program_path
Following almost the same format as the other two add functions for item name and description, this function first presents the user with an option to return to the previous menu, enter the full path manually, or use a file chooser dialog window.

In a similar fashion to edit_startup_item_program_path, if a file chooser dialog window is selected, the filedialog function from tkinter is called to present a file dialog window so the user can graphically navigate to the executable.

This add function takes one parameter, which is the name of the startup item, in case, when creating a new startup item, a name is set first before choosing the path. The parameter is optional, and if no name is given, a default of "Startup Item" is used.

There is a loop added similar to the other two add functions, but this loop only checks to see that the user entered something, if manually entering a path, or choose something, if using the file dialog window. It doesn't validate the input to make sure it's a proper and full path, etc.