Minor changes to the startup_edit module
Modified the module to import the JsonSchemaKeys enum class and use its members in the edit_startup_item function for references to specific JSON object members.

Tested that it worked and also noted that the function edit_startup_item_arguments_list is the only function left that needs to be refactored to use the updated user_menu_chooser function.