Minor changes to the existing function edit_file_chooser
Renamed it to existing_file_chooser to better reflect its usage. Also, updated the call to it from the function edit_startup_item_program_path.

Next, added a comment before line 110 (now 112) to explain the call to the tkiner filedialog.askopenfilename function.

Finally, changed the path separators used for the initialdir parameter to be Windows with escape characters, so "\\" instead of "/". Both work for the tkinter function, but aesthetically, it looks more in line with this being a Windows tool.