Refactored the rest of the code to use the new user_menu_chooser function
Updated the function declaration and docstring for the function edit_startup_item_arguments_list to not take the second parameter arg_count, since it's no longer needed. Also, updated the call from the function edit_startup_item to not pass in that paremeter.

Within the function edit_startup_item_arguments_list, refactored things to grab the length of the arguments list passed in to track the argument count. Also, rewrote the menu sections, both for the first items menu, with the options to edit each argument, add an argument, or delete an argument, and the delete items menu, with the options to delete each existing argument.

For both menus, removed any strings formatted to include a menu option number and all newlines, so the strings passed in for each menu option are in the correct format used by the new user_menu_chooser function. Also, removed any calls to user_menu_chooser that passed in a string, and made sure only a list of strings was passed in.

Next, removed variables no longer needed such as menu_header and total_menu_choices. Also, fixed a call to the add_startup_item_arguments list when the user chooses to add an argument. The call was passing in the original argument list passed, not the local copy of the updated argument list. Now it passes in the local copy.

Finally, tested things out. As part of the testing, adjusted the startup_data.json file and saved it, but then made a copy of the saved file called startup_data--issue25.json and put it in the testing folder. Also, as part of the testing, created a copy of startup.ps1 from /devenv/ to the root of the feature branch: /devenv/feature_addons/startup_data_modifier_tool. Modified the feature branch copy to look for startup_data.json in the same directory.

This way, further testing of changes to the startup_data.json file can be checked to make sure the startup PowerShell script still works. In this case, the changes saved in startup_data--issue25.json were tested and opened correctly by the startup PowerShell script.