Added the function defintion for add_startup_item
Created a function definition for add_startup_item. Updated the docstring and commented out code at the very end of the module.

Still need to code in the function. It takes no arguments but will return a fully formed startup item dictionary. It will also validate the startup item against the Startup Item schema before returning, just to be on the safe side.