Worked some more on the function add_startup_item
In add_startup_item, added in code to create a new Python dictionary and to then populate it with all the keys for a startup item with default values of None. Also created the main loop for the function and added in calls to set the name, description, file path, and any arguments.

Still need to add a bit more functionality, such as calculating and setting the ArgumentCount parameter, setting the Browser parameter, and finally, the ItemNumber parameter.