Updated add_startup_item
Added in the code to fill out the startup item parameters "ItemNumber", "Browser", and "ArgumentCount". This gets filled in when the user chooses to save the startup item.

Still need to code in the actual saving of the startup item. Also, while testing, realized that before saving, the startup item needs to be validated, both to make sure it meets the JSON Startup Item Schema, and also to make sure that if a user doesn't fill out each section of the startup item, the default None isn't used.

One possibility is to have the new_item dictionary not populated with None when adding in the keys, but actual defaults (such as 0, blank strings, etc.).

Another possibility is to add a menu choice that goes through setting up the whole startup item, and then refactor the existing menu choices to allow the user to edit what they set. This can get rid of the confirming that happens in add_startup_item_name and add_startup_item_description. There should also probably be a menu option to view the startup item.