Renamed some module imports
Realized that in the jsonfn module, the modules startup_add and startup_edit were imported as the variables deps_start_add and deps_start_edit, respectively.

However, that wasn't consistent with what's listed in the import_lists.txt file in the dependencies folder. Updated all references in module jsonfn to make it consistent.