Refactored add_startup_item
Adjusted the building of new_item to use default values for each property, such as 0 or "", depending on the property type.

Also, changed the menu so there are now 8 choices. Added an option as the first menu choice that lets the user create the actual startup item.

That is, rather than the user needing to select several menu options, one for each property, this first option goes through the process of setting up each property.

The rest of the menu choices allow for the user to adjust the Name, Description, FilePath, and ArgumentList properties, to view the startup item, to save the startup item, and to return to the previous menu.

These options haven't been coded in yet except for the last option, but the menu displays them.