Refactored other aspects of the module startup_add
Created a new helper function called startup_item_setup, which takes a reference to a dictionary, and calls the functions to add a Name, Description, FilePath, and ArgumentList.

The first menu option in the main menu calls this helper function. That way, it will be easier to make changes if need be. This also makes the code for the add_startup_item function less confusing.