Made more changes to the module startup_add
Refactored the add functions for Name and Description by removing the menus within those functions, and simply keeping the user input and return lines.

For the add function for FilePath, removed the menu option to return without setting a file path (for now). As a result, the user can only choose to enter the path manually or use the file chooser dialog.

For the add function for ArgumentList, made a slight adjustment so the argument number is listed when prompting the user to pass in the new argument.