Minor adjustment to module startup_add
Added an import statement to import the module startup_edit.

This is for the menu options in the function add_startup_item that allow the user to adjust the properties Name, Description, FilePath, and ArgumentList after they've set them up initially. The plan is to have those menu options call the necessary functions in the module startup_edit since they already exist.