Readded use of the is_prod variable
Decided to reuse the existing global variable is_prod and the function is_production in the helper module. Updated the comments for the variable to better explain its purpose.

Updated the function get_prod_path to check the environment and if it's no a production environment, to return a list containing all the subfolders from CompStart - the project root folder - to the correct config location.

This was necessary because the os.getcwd function would return the absolute path from C to the CompStart folder, both when tested using VS Code and when tested using Windows Terminal. Unsure why this is the case, but will research this later.