Started working to validate the read-in JSON data
Per the notes from issue #20, there are 3 scenarios currently to check for:

1. A completely blank JSON file
2. An empty JSON object
3. JSON data within a JSON object, but it's not valid startup data

For the first scenario, the try-except block handles this. Specifically, a JSONDecodeError exception is thrown and printed out using the prettify_io_error function in the pretty module. Execution then returns right back to the main loop. For the other two scenarios, added two checks on the JSON data that's read in the try block.

If the JSON file just has an empty object, a custom error is printed out stating "JSON data is blank" using the prettify_custom_error function in the pretty module. The rest of the json_reader function then finishes and execution returns to the main loop. This is checked by seeing if the length of the json_data list variable is 0. If the JSON file has data, but it's not valid startup data, a custom error is printed out stating "JSON data isn't valid startup data" using the prettify_custom_error function in the pretty module. The data is checked by using the json_data_validator function in the helper module.

Finally, if both checks are passed, then the read_json_success variable is set to True since it's set to False as default. Also, had to change the default value of the return_message variable since it would print out "Startup data read in successfully!" even when an error message was printed. Changed the default message to "There was a problem reading in the startup data!\nPlease see the error details above." and only when both checks above are passed is the message changed to "Startup data read in successfully!".