Fixed the loop issue during JSON data validation
An optional boolean argument was added to the json_reader function declaration. The function docstring explains it fully:

This optional argument is a variable specifying if the JSON file that will be read is going to be one of the JSON schema files. This was added because in order to confirm that read in JSON startup data is valid, the function json_data_validator from the helper module is called. However, that function then calls this function, which creates a loop.

This variable will be used specifically to avoid that situation. The default is False, so most existing calls to this function will still work. The call in the json_data_validator function will pass in a value of True for this variable.

In the check for valid JSON data, this boolean argument is checked. If the value is True, then the check is skipped which avoids the loop.