Refactored the function new_file_chooser
In the module chooser, refactored the new_file_chooser function. Changed the two boolean variables from is_default and to_continue to is_default and create_file. Updated the docstring to explain what a value of True means for each variable and what False means. Reworded the third menu choice to specify that returning to the main menu means skipping creating a startup file. Finally, refactored the if block as necessary.

Also made the necessary variable name changes in comp_start to stay in line with the new variable names in new_file_chooser.