Handled edge case for deleting starting items
While testing, realized that when editing a single item, if all the items are deleted one by one, and the delete item option is picked, the program continued by asking which item number to delete but would show "[1-0]" meaning from item number 1 to item number 0. This was because the TotalItems JSON field would have the correct value of 0.

This was adjusted so that if this case occurs, the following message is printed out:

There are no items to delete! Please add a new startup item first...