Adjusted some error printing formatting
In the function json_reader, when there's an exception caught while trying to read in the JSON data, the function prettify_io_error from the pretty module is called. However, this function just returns a formatted error message and lets the caller function print it out.

This was causing some inconsistent error printing as the function prettify_custom_error, which was called when the JSON data was read in but was blank or not validated, would print out the error passed in. Additionally, in the case of the function json_reader, the returned and formatted error message would be then passed back up the stack as the return_message variable.

This was changed to take the return of the prettify_io_error function and feed it to the prettify_custom_error function. Now all 3 different errors are printed out in the same format. Additionally, the return message from the json_reader function stays as the default of stating there was as problem reading in the data, unless the data read in was successful, not blank, and properly validated.

Lastly, removed a "\n" from one of the print lines in the main loop to keep spacing looking the when the same error message is printed out regardless of which main loop menu choice was chosen.