Multiple changes to how errors are handled
Refactored several functions to print any error messages directly, such as the functions json_creator, json_writer, and json_reader. These functions now print any error messages that occurred during their execution prior to returning up the stack.

Adjusted the prettify_io_error function slightly so it's more clear that the error is a Python system error instead of an error message coded in by the programmer.

Overall, set things up so that each function that could have an issue, whether through a Python Exception or another type of error, prints out the error message and the function name it occurred in as execution goes along. This creates a chain of printed-out error messages, which allows one to trace back, similar to a Python error traceback.

Tested this out by setting the default_startup.json file to be completely blank, to be an empty JSON object, and to be invalid startup JSON data. In all 3 cases, an error is printed starting from the json_creator function > to the generate_default_startup_data function > to the json_reader function > to the json_data_validator function when applicable.