Slight adjustment to function json_writer
While testing, noted the following:

From the main loop, if the user chose Option 2 - to "Create a new startup file" - then Option 2 - to "Create a new startup file with programs that you choose" - then N for the question "Would you like to completely overwrite this file [Y/N]", the return message of "Skipped writing startup file" was printed out but so was an error message with the same wording and stating the function as json_writer.

The reason for this is in lines 161-163, where the code to print out any errors was simply checking the boolean variable write_json_success. However, since the variable return_message is used to return messages to print even when there was no actual JSON data to write, such as when skipping writing due to user choice, the if-block had to be changed.

In this case, whenever there is a message to return and print that's no an error, the variable file_mode will not be blank. So, the if-block was updated to check for both the boolean variable write_json_success and the string variable file_mode.