Taken from the comments for issue #31
While starting to work on the second task, realised that because the branch for this issue is based on the branch features/startup_data_modifier_tool, certain newer code changes made in the branch dEhiN/issue4 will need to be merged in. Specifically, in the other branch, the function is_production was moved from the main comp_start.py script to the module helper.

Extrapolating the location of the default startup data JSON file, or even the in-use startup data JSON file, to be stored in only one place should then be placed in the same module for consistency. However, the branch for this issue doesn't have that code change, so it would be difficult to make the same change.

Will move the 2 tasks left to a new enhancement issue for later and will resolve this issue.