Renamed schema folder
During testing, ran into an error while trying to display the startup file data. There was a Python FileNotFound error when trying to validate the startup data because the directory structure is supposed to be /config/schema for the schema files.

However, the folder was named schemas instead. Unsure if this is from a merge conflict that was overlooked during PR #37 but changed the folder name to schema to fit in line with the code.