Fixed blank errror message printing edge case
During the previous test, while debugging the issue, realized that the error handling code in the function json_reader causes a blank error message to be printed out when there's a Python Exception that's printed out.

The function was refactored to check for the value of the boolean variable read_json_success just prior to the return statement, and if it was False, print out the string variable return_message as a custom error.

The boolean variable was set to False and the string variable was set to blank as defaults. This was done so that custom error messages could be supplied for situations where the JSON file existed and was read in - so no Python Exceptions - but the data was incorrect or invalid.

However, in the try-except block where the actual loading of loading of the JSON data from the file is done, the except section catches any Python Exception errors. It then call prints out the error message.