Refactored error handling functionality
Realized that in the function json_data_validator, while any Python Exceptions caught when trying to validate the JSON data is printed, nothing is printed if there's a problem while trying to read the JSON schema files.

Specifically, an error from json_reader will print out specifying the Python Exception (such as FileNotFoundError), and then the next error printed will be also from json_reader stating:

JSON data isn't valid startup data

Since this is misleading, the function json_data_validator was refactored to print out a message in this scenario.

Further changes were also made to the above error message. That error message is shown when, in json_reader, the JSON data is successfully read in but fails data validation. However, since the function json_data_validator could return False (meaning the data validation failed) for reasons other than the actual JSON data failed to validate against the JSON schema, that error message was updated to better reflect the state of execution at that point.

Finally, changes were made to both json_reader and json_creator so that when they return to the main loop, neither of them prints out any final message to the user to check the error messages above. Rather, in the main loop for both case 2 and 3, if the status_state variable that's returned from the jsonfn functions is false, indicating the jsonfn function failed, the variable status_message is updated to the final error message alerting the user to check the other error messages. This is then printed out.

This final error message was also extrapolated to a global variable so if it needs to be changed, it can be done so in only one place.