Brought back some old testing files
After PR #37, several testing related files in the testing folder were deleted. I wanted to keep them for posterity but was unable to during the PR and subsequent merge.

I went back to commit 6da619b when the files still existed, copied them out of the project folder, and copied them back to the head of dEhiN/issue4.

However, they are all stored in a sub-folder of testing called old-previously-deleted-testing.