Updated references to comp_start
Some of the modules imported the main script. They referenced it as comp_start. However, since the script was renamed to CompStart, those references had to be updated.

Also changed around the welcome blurb shown when first choosing to add an item.

The rest of the changes are minor changes where, due to the line length setting in VS Code, some long lines were broken up.