Worked on functionality for adding an item
After reviewing the code, it was realized that the functions to adjust each property of the startup item already existed.

So, options 2 to 5 were filled in on the add a startup item menu. Each option just calls the equivalent function starting with "add_" and updates the variable new_item, which stores the startup item data.

Note: If someone skips option 1, options 2-5 will still work.