Comment changes
While creating the README for the experimental_content folder, both Python scripts in the tkinter/tkinter_examples folder were reviewed to recall what they did.

While reviewing, realized that the comments for the hello_world.py file was a little confusing. Redid the comments to have one fully explanatory comment at the beginning of the script.