Removed an import statement
Realized that the module startup_edit was being imported because it was believed at one point that the functions in that module would be used.

However, since they are not used even for the options to adjust/edit the startup item properties, there is no need for the import statement.