Fix the ItemNumber bug
As mentioned in the comments for issue #4, after a new startup item is created, the property ItemNumber isn't changed.

Fixed that so now the propery has the correct value based on the startup data JSON file.

Created two functions in the module helper to assist with this. The first is called get_count_total_items and, like the name implies, returns just the count of the total items currently in the startup data file.

The second function is called get_startup_filename and takes a boolean argument. This function returns the JSON filename depending on the boolean argument. See the function docstring for more details.

Note: This second function can help in issue #36.