Added checks in match-case block
In the match-case block in the function add_startup_item, added an if-else block to check if the user has created a startup item first.

Choices 2-5 - to edit the startup item properties - are now collapsed into a single case block. Within it, the ItemNumber property is checked and if it's still 0, which is what it's initialized at but then updated when the user chooses choice 1 - to create a new startup item - a message is printed out to create the startup item first.

Still need to finish the logic for the else block.