Added a case block to the main menu for quitting
While testing, realized that the main menu match-case block doesn't have a case-block for option 5 - to quit the program.

This is because the function user_menu_chooser in the module chooser adds an option to quit the program by default. If the user chooses this option, the command sys.exit is used.

However, as mentioned in the comments for the case 5 block, just in case execution returns to the main menu (ex. perhaps through a future bug), this block will now ensure the user can quit the program.