Finished work referenced in previous commit
In commit b5e2259, several of the menu options in the add a startup item menu are collapses into a single case statement.

This was done to cut down on the code that checks if the user has gone through option 1 first via an if-block

The else part of the if-block is now filled in by using an if-block on the variable user_choice, which is also what the match-case uses.

Although an if-block with 2 elif sections and an else section had to be added, thereby nesting two if-blocks, it still overall is cleaner code.