Updated all references
Went through all the Python files and updated any code where a variable stores the name of one of the JSON files.

All the locations that need this file name now call the function get_startup_filename in the helper module.

Please note that this is not the case for the Schema JSON files. However, those files are referenced in only one line of code in the function json_data_validator, so there is no need to refactor them to a helper function.