Refactored the loop for adding a new startup item
After adding the option to view the startup item, it was noted that refactoring of the whole loop could be done to be more streamlined.

After the user chooses a menu option, an if-block is used to check if the user chose to quit or if the user chose an option other than 1 but hasn't gone through option 1 yet.

If both those checks pass, then the match-case is called.