Started working on option 7
Started working on the case to save the startup item in the Add a Startup Item menu.

After the code was added, when testing, an IndexError occurred. The existing startup data was read in without any issues, so will need to go through this to see what the problem is.

Referencing issue #4 on GitHub for visibility.