Added option 7 to save a startup item
Based on the comments for issue #4, it was decided to resolve the IndexError issue mentioned in commit 6132373 by creating 2 save_startup_item functions.

The existing code was copied to the startup_edit module where the function was renamed to save_modified_startup_item. While the function was changed, and the call to it from the Edit a Startup Item menu updated, the function code wasn't edited. Therefore, while this hasn't been tested, the function should still work.

The function in the startup_add module was renamed to save_new_startup_item. The parameter names were updated as well and the docstring. Finally, the code was modified as necessary.

This was then tested and a fourth startup item was added to the startup data file.